amd64 (and i386): packet pointer, session pointer and mailbox pointer now stored as longs, converted to-from jlongs.
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@15920 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ jnetcap/impl/com/untangle/jnetcap/PacketMailbox.java
@@ -34,5 +34,5 @@ public interface PacketMailbox
     public Packet read( int timeout );
 
     /* Retrieve the value of the C pointer */
-    public int pointer();
+    public long pointer();
 }
